+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - [GUI] Proportional, granular progress updates for installing (#4055 by: HebaruSan)
 - [GUI] Modpack compatibility prompt, GameComparator clean-up (#4056 by: HebaruSan)
 - [ConsoleUI] Add downloads column for ConsoleUI (#4063 by: HebaruSan)
+- [ConsoleUI] Play game option for ConsoleUI (#4064 by: HebaruSan)
 
 ### Bugfixes